Update alternative.md to add solution for ARM QNAP NAS
The guide for installing on QNAP systems sometimes doesn't work for ARM QNAP systems, eg TS-233: instead of a container which starts, it gets stuck in a boot loop including the log message "unable to iopause: Function not implemented". An alternative approach which did work for this user is the "docker compose" approach (QNAP Container Station UI: "Applications\Create"). This pull request adds a corresponding hint to the documentation.